+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Version History
 
+## 0.0.0.5
+
+* Fix multiquery decorator
+* Fix DumperLogger show version
+* Fix chunk_query parts split
+* Fix freeze to_reader() method where it have post queryes
+* Change hidden methods to protected methods
+* Rename AbstractReader to ExampleReader
+* Update READEME.md
+
 ## 0.0.0.4
 
 * Refactor BaseDumper